rename source files bmc_solver.py -> BlochMcConnellSolver.py bmc_tool.py -> BMCTool.py
