+++ src/bmctool/BlochMcConnellSolver.py
